requirements.txt
google-genai
langchain
langchain-google-genai
python-dotenv
rich
pytest
streamlit
